# A benchmark for our trigonometric functions and a few performance improvements