--- a/numerics/sin_cos.hpp
+++ b/numerics/sin_cos.hpp
@@ -5,7 +5,15 @@ namespace numerics {
 namespace _sin_cos {
 namespace internal {
 
+#define PRINCIPIA_INLINE_SIN_COS 0
+
+#if PRINCIPIA_INLINE_SIN_COS
+inline
+#endif
 double __cdecl Sin(double x);
+#if PRINCIPIA_INLINE_SIN_COS
+inline
+#endif
 double __cdecl Cos(double x);
 
 }  // namespace internal
@@ -16,3 +24,7 @@ using internal::Sin;
 }  // namespace _sin_cos
 }  // namespace numerics
 }  // namespace principia
+
+#if PRINCIPIA_INLINE_SIN_COS
+#include "numerics/sin_cos.cpp"
+#endif